+++ packages/nimble-blazor/NimbleBlazor.Components/Appearance.cs
@@ -10,8 +10,7 @@ public enum Appearance
 
 internal static class AppearanceExtensions
 {
-    private static readonly Dictionary<Appearance, string> _appearanceValues =
-        Enum.GetValues<Appearance>().ToDictionary(id => id, id => Enum.GetName(id)!.ToLowerInvariant());
+    private static readonly Dictionary<Appearance, string> _appearanceValues = AttributeHelpers.GetEnumNamesAsKebabCaseValues<Appearance>();
 
     public static string? ToAttributeValue(this Appearance? value) => value == null ? null : _appearanceValues[value.Value];
 }