+++ packages/nimble-components/src/all-components.ts
@@ -1,0 +1,28 @@
+/**
+ * Import of all the web components available in Nimble.
+ * Production applications are encouraged to import only components
+ * that are required instead of leveraging this file.
+ */
+
+import './breadcrumb';
+import './breadcrumb-item';
+import './button';
+import './checkbox';
+import './drawer';
+import './icons/all-icons';
+import './listbox-option';
+import './menu';
+import './menu-item';
+import './number-field';
+import './select';
+import './switch';
+import './tab';
+import './tab-panel';
+import './tabs';
+import './tabs-toolbar';
+import './text-area';
+import './text-field';
+import './theme-provider';
+import './toggle-button';
+import './tree-item';
+import './tree-view';