+++ packages/nimble-components/src/tab/index.ts
@@ -5,8 +5,6 @@ import {
 } from '@microsoft/fast-foundation';
 import { styles } from './styles';
 
-export type { Tab };
-
 declare global {
     interface HTMLElementTagNameMap {
         'nimble-tab': Tab;
@@ -16,7 +14,7 @@ declare global {
 /**
  * A nimble-styled HTML tab
  */
-class Tab extends FoundationTab {}
+export class Tab extends FoundationTab {}
 
 const nimbleTab = Tab.compose({
     baseName: 'tab',