+++ packages/nimble-components/src/tabs/tests/tabs.spec.ts
@@ -2,13 +2,14 @@ import {
     DesignSystem,
     Tabs as FoundationTabs
 } from '@microsoft/fast-foundation';
-import { html } from '@microsoft/fast-element';
-import '..';
+import { Tabs } from '..';
 
 describe('Tabs', () => {
     it('should have its tag returned by tagFor(FoundationTabs)', () => {
-        expect(html`${DesignSystem.tagFor(FoundationTabs)}`.html).toBe(
-            'nimble-tabs'
-        );
+        expect(DesignSystem.tagFor(FoundationTabs)).toBe('nimble-tabs');
+    });
+
+    it('can construct an element instance', () => {
+        expect(document.createElement('nimble-tabs')).toBeInstanceOf(Tabs);
     });
 });