+++ packages/nimble-components/src/text-field/tests/text-field.spec.ts
@@ -2,13 +2,18 @@ import {
     DesignSystem,
     TextField as FoundationTextField
 } from '@microsoft/fast-foundation';
-import { html } from '@microsoft/fast-element';
-import '..';
+import { TextField } from '..';
 
 describe('TextField', () => {
     it('should have its tag returned by tagFor(FoundationTextField)', () => {
-        expect(html`${DesignSystem.tagFor(FoundationTextField)}`.html).toBe(
+        expect(DesignSystem.tagFor(FoundationTextField)).toBe(
             'nimble-text-field'
+        );
+    });
+
+    it('can construct an element instance', () => {
+        expect(document.createElement('nimble-text-field')).toBeInstanceOf(
+            TextField
         );
     });
 });